fix: 修复 request_id 超过 int 的 json 解析错误

+++ conf/conf.go
@@ -1,28 +1,28 @@
 package conf
 
 type PcsResponseBase struct {
-	ErrorCode int  	 `json:"error_code"`
-	ErrorMsg  string `json:"error_msg"`
-	RequestID int `json:"request_id"`
+	ErrorCode int     `json:"error_code"`
+	ErrorMsg  string  `json:"error_msg"`
+	RequestID float64 `json:"request_id"`
 }
 
 type CloudDiskResponseBase struct {
-	ErrorCode int  	 `json:"errno"`
-	ErrorMsg  string `json:"errmsg"`
-	RequestID int `json:"request_id"`
+	ErrorCode int     `json:"errno"`
+	ErrorMsg  string  `json:"errmsg"`
+	RequestID float64 `json:"request_id"`
 }
 
 type TestDataConfig struct {
-	ClientID string
-	ClientSecret string
-	RedirectUri string
-	Code string
-	AccessToken string
-	RefreshToken string
-	Dir string
-	FsID uint64
-	Path string
-	LocalFilePath string
+	ClientID        string
+	ClientSecret    string
+	RedirectUri     string
+	Code            string
+	AccessToken     string
+	RefreshToken    string
+	Dir             string
+	FsID            uint64
+	Path            string
+	LocalFilePath   string
 	TranscodingType string
 }
 
@@ -30,8 +30,8 @@ const Version = "0.0.5"
 
 const (
 	BaiduOpenApiDomain = "https://openapi.baidu.com"
-	OpenApiDomain = "https://pan.baidu.com"
-	PcsDataDomain = "https://d.pcs.baidu.com"
+	OpenApiDomain      = "https://pan.baidu.com"
+	PcsDataDomain      = "https://d.pcs.baidu.com"
 )
 
 // 测试参数
